Refactor CSV File Output Mechanism

+++ Models/Flight.cs
@@ -5,7 +5,7 @@ namespace Airport.Models {
     public class Flight {
         public required string FlightID { get; set; } // probs not practical for large datasets? TODO -> look for an alternative way.
         // check booking TODO
-        public required int Cost { get; set; } 
+        public required decimal Cost { get; set; } 
 
         public required string DepartedCountry { get; set; }
         public required string DestinatedCountry { get; set; }
